changed to numbers

+++ src/main/java/rocks/zipcode/atm/CashMachine.java
@@ -43,7 +43,7 @@ public class CashMachine {
     public String getAccoutEmail() {
         return accountData.getEmail();
     }
-    public int getAccoutBal() {
+    public double getAccoutBal() {
         return accountData.getBalance();
     }
 
@@ -63,7 +63,7 @@ public class CashMachine {
         );
     }
 
-    public void deposit(int amount) {
+    public void deposit(double amount) {
         if (accountData != null) {
             tryCall(
                     () -> bank.deposit(accountData, amount),
@@ -72,7 +72,7 @@ public class CashMachine {
         }
     }
 
-    public void withdraw(int amount) {
+    public void withdraw(double amount) {
         if (accountData != null) {
             tryCall(
                     () -> bank.withdraw(accountData, amount),
